+++ Assets/NatShare/Plugins/Managed/INatShare.cs
@@ -8,11 +8,10 @@ namespace NatShareU {
 	using UnityEngine;
 
 	public interface INatShare {
-		bool ShareText (string text, ShareCallback callback);
-		bool ShareImage (byte[] pngData, string message, ShareCallback callback);
-		bool ShareMedia (string path, string message, ShareCallback callback);
-		bool SaveToCameraRoll (byte[] pngData);
-        bool SaveToCameraRoll (string videoPath);
+		bool Share (byte[] pngData, ShareCallback callback);
+		bool Share (string media, ShareCallback callback);
+		bool SaveToCameraRoll (byte[] pngData, string album);
+        bool SaveToCameraRoll (string path, string album, bool copy);
         Texture2D GetThumbnail (string videoPath, float time);
 	}
 }