+++ Assets/NatShare/Plugins/Managed/NatShare.cs
@@ -15,49 +15,33 @@ namespace NatShareU {
 		#region --Client API--
 
 		/// <summary>
-        /// Share plain text with the native sharing UI.
-        /// Returns true if the text can be shared.
-        /// </summary>
-        /// <param name="text">Text to be shared</param>
-		/// <param name="callback">Optional. Callback to be invoked once sharing is complete</param>
-		public static bool ShareText (string text, ShareCallback callback = null) {
-			if (string.IsNullOrEmpty(text)) {
-				Debug.LogError("NatShare Error: Text being shared is null");
-				return false;
-			}
-			return Implementation.ShareText(text, callback);
-		}
-
-		/// <summary>
         /// Share an texture with the native sharing UI.
         /// Returns true if the image can be shared.
         /// </summary>
         /// <param name="image">Image to be shared</param>
-		/// <param name="message">Optional. Message to be shared with image</param>
 		/// <param name="callback">Optional. Callback to be invoked once sharing is complete</param>
         [Doc(@"ShareImage")]
-		public static bool ShareImage (Texture2D image, string message = "", ShareCallback callback = null) {
+		public static bool Share (Texture2D image, ShareCallback callback = null) {
 			if (!image) {
 				Debug.LogError("NatShare Error: Texture being shared is null");
 				return false;
 			}
-			return Implementation.ShareImage(image.EncodeToPNG(), message, callback);
+			return Implementation.Share(image.EncodeToPNG(), callback);
 		}
 
 		/// <summary>
-        /// Share a media file with the native sharing UI.
-        /// Returns true if media file is found and can be shared.
+        /// Share plain text with the native sharing UI.
+        /// Returns true if the text can be shared.
         /// </summary>
-        /// <param name="path">Path to media file</param>
-		/// <param name="message">Optional. Message to be shared with image</param>
+        /// <param name="media">Media to be shared</param>
 		/// <param name="callback">Optional. Callback to be invoked once sharing is complete</param>
-        [Doc(@"ShareMedia")]
-		public static bool ShareMedia (string path, string message = "", ShareCallback callback = null) {
-			if (string.IsNullOrEmpty(path)) {
-				Debug.LogError("NatShare Error: Path to media file is invalid");
+		[Doc(@"ShareMedia")]
+		public static bool Share (string media, ShareCallback callback = null) {
+			if (string.IsNullOrEmpty(media)) {
+				Debug.LogError("NatShare Error: Media being shared is null or empty");
 				return false;
 			}
-			return Implementation.ShareMedia(path, message, callback);
+			return Implementation.Share(media, callback);
 		}
 
 		/// <summary>
@@ -65,13 +49,14 @@ namespace NatShareU {
         /// Returns true if the image can be saved to the camera roll.
         /// </summary>
         /// <param name="image">Image to be saved</param>
+		/// <param name="album">Optional. Album where to save image</param>
         [Doc(@"SaveToCameraRoll")]
-		public static bool SaveToCameraRoll (Texture2D image) {
+		public static bool SaveToCameraRoll (Texture2D image, string album = "") {
 			if (!image) {
 				Debug.LogError("NatShare Error: Texture being saved is null");
 				return false;
 			}
-			return Implementation.SaveToCameraRoll(image.EncodeToPNG());
+			return Implementation.SaveToCameraRoll(image.EncodeToPNG(), album ?? "");
 		}
 
 		/// <summary>
@@ -79,13 +64,15 @@ namespace NatShareU {
         /// Returns true if the file is found and can be saved to the camera roll.
         /// </summary>
         /// <param name="path">Path to media file</param>
+		/// <param name="album">Optional. Album where to save image</param>
+		/// <param name="copy">Optional. Should file be copied or moved to the camera roll?</param>
         [Doc(@"SaveToCameraRoll")]
-		public static bool SaveToCameraRoll (string path) {
+		public static bool SaveToCameraRoll (string path, string album = "", bool copy = true) {
 			if (string.IsNullOrEmpty(path)) {
 				Debug.LogError("NatShare Error: Path to media file is invalid");
 				return false;
 			}
-			return Implementation.SaveToCameraRoll(path);
+			return Implementation.SaveToCameraRoll(path, album ?? "", copy);
 		}
 
 		/// <summary>
@@ -93,7 +80,7 @@ namespace NatShareU {
         /// If the thumbnail cannot be loaded, the function will return `null`.
         /// </summary>
         /// <param name="videoPath">Path to recorded video</param>
-        /// <param name="time">Optional: Time to get thumbnail from in video</param>
+        /// <param name="time">Optional. Time to get thumbnail from in video</param>
         [Doc(@"GetThumbnail", @"GetThumbnailDiscussion"), Code(@"Thumbnail")]
 		public static Texture2D GetThumbnail (string videoPath, float time = 0f) {
 			if (string.IsNullOrEmpty(videoPath)) {